rm extra files
